refactor OCRActivity logic
Now MLKitOcrScanActivity will be in the public library and use BaseOcrScanActivity, in the private we will have 2 new activities HybridOcrScanActivity and TesseractOcrScanActivity
