+++ lib/src/main/res/layout/activity_ocr_scan.xml
@@ -4,4 +4,4 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.OcrScanActivity" />
+    tools:context=".ui.BaseOcrScanActivity" />
