+++ lib/src/main/res/layout/fragment_ocr_scan.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.OcrScanActivity">
+    tools:context=".ui.BaseOcrScanActivity">
 
     <androidx.camera.view.PreviewView
         android:id="@+id/preview_view"
